Fix thread-safety issues with task-related environment functions
