+++ djangae/contrib/backup/tests/test_views.py
@@ -5,7 +5,7 @@ from django.test import (
 
 from djangae.contrib import sleuth
 from djangae.test import TestCase
-from djangae.decorators import _TASK_NAME_HEADER
+from djangae.tasks.middleware import _TASK_NAME_HEADER
 from ..views import create_datastore_backup
 
 
@@ -24,13 +24,14 @@ class DatastoreBackupViewTestCase(TestCase):
     def test_get_params_propogate(self):
         request = RequestFactory().get('/?kind=django_admin_log&bucket=foobar')
         request.META[_TASK_NAME_HEADER] = "test"
-        with sleuth.Fake('djangae.contrib.backup.views.backup_datastore', None) as backup_fn:
-            create_datastore_backup(request)
-            self.assertTrue(backup_fn.called)
-            self.assertEqual(
-                backup_fn.calls[0][1],
-                {
-                    'bucket': 'foobar',
-                    'kinds': [u'django_admin_log']
-                }
-            )
+        with sleuth.fake("djangae.tasks.decorators.is_in_task", True):
+            with sleuth.fake('djangae.contrib.backup.views.backup_datastore', None) as backup_fn:
+                create_datastore_backup(request)
+                self.assertTrue(backup_fn.called)
+                self.assertEqual(
+                    backup_fn.calls[0][1],
+                    {
+                        'bucket': 'foobar',
+                        'kinds': [u'django_admin_log']
+                    }
+                )
