+++ djangae/contrib/backup/views.py
@@ -1,7 +1,7 @@
 import logging
 
 from django.http import HttpResponse
-from djangae.decorators import task_only
+from djangae.tasks.decorators import task_only
 
 from .tasks import backup_datastore
 from .utils import get_backup_setting
